backend/evaluation/requirements.txt
# Core dependencies
pyyaml>=6.0.1
aiohttp>=3.9.0
asyncio>=3.4.3

# LLM providers
openai>=1.12.0
anthropic>=0.18.0
google-genai>=0.3.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Graph processing
networkx>=3.2.1

# NLP & Embedding
sentence-transformers>=2.3.0
scikit-learn>=1.3.0

# Visualization (optional)
matplotlib>=3.8.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
